Spawn planks and rope coil
**Why?**
Just to get some more things in there. We'll have a bear sitting on the
planks of wood that gives the user the prompt to get the sugar and
whatnot. The rope is a little further off because it'll be in another
room and be *combined* with the planks to make a bridge. First place
here we come.