Extract debugging plugin
## This Commit

Extracts a plugin that encapsulates the logic needed to allow developers
to pass levels when speech isn't working (either because they compiled
without DG or DG is flaky). It also can be useful for those cannot use
the existing speech feature due to accessibility restrictions.

## Why?

Because these are separate/transparent to the rest of the app/systems
thanks to the use of the `EventWriter<SpeechEvent>` and can be scuttled
away to clean some things up.