Use production Deepgram for WS transcription
## This Commit

Updates the repo to use the official production Deepgram endpoint for WS
transcription instead of the proxy.

## Why?

Because that is more correct and stable. Then we can take down the proxy
and stop managing it.

## Notes

This requires those building the repo to compile a token into the binary
which will be available on the client. We think this is fine for now and
can come up with something else if we want later.

Co-authored-by: Nikola Whallon <nikola@deepgram.com>